Move MatrixMap to matrix_map.rs
