+++ src/main.rs
@@ -5,6 +5,7 @@ mod matrix_state;
 mod path_map;
 mod plugin_iterator;
 mod plugin_logs;
+mod matrix_map;
 
 use crate::clargs::{MatricksArgs, MatricksSubcommand};
 use crate::core::matricks_core;
